Eliminate circular deps
Eliminatin circular deps almost always yields some nice architectural
wins and this is no exception. DagsterEnv (which began as InputManager)
is gone! All rejoice. Things are much cleaner now.
