+++ python_modules/dagster/dagster/cli/pipeline.py
@@ -8,9 +8,9 @@ import yaml
 
 import dagster
 from dagster import check
+from dagster.core.definitions import ExecutionGraph
 from dagster.core.execution import execute_pipeline_iterator
 from dagster.core.errors import DagsterExecutionFailureReason
-from dagster.core.graph import ExecutionGraph
 from dagster.graphviz import build_graphviz_graph
 from dagster.utils.indenting_printer import IndentingPrinter
